requirement.txt
easydict
mxnet-cu102
Pillow
sklearn
torch
torchvision
tqdm
matplotlib
